refactor(125/WS1.4): extract state.rs (config, handle, AppState) (#125)
Second move in the src/server/mod.rs decomposition series (WS1.4).

WHAT:
- Extract src/server/state.rs (565 lines): ServeOptions, ServerHandle
  (+impls + Drop), DaemonConfig (+impls + Default), DaemonUpdateConfig
  (+Default), DEFAULT_QUIC_PORT, the config default_* helpers +
  shared_cache_dir, AppState, CachedUpgradeCheck. All moved verbatim.
- mod.rs: -547 lines (25740 → 25217). Private `mod state;` with a
  `pub use state::{DaemonConfig, ServeOptions, ServerHandle, DEFAULT_QUIC_PORT,
  default_*}` re-export preserving the public x0x::server::* API surface.
  Internal types (AppState, DaemonUpdateConfig, CachedUpgradeCheck) imported
  privately via `use state::{...}` — they are NOT publicly re-exported.

VISIBILITY (internal to `server`, no behavior change):
- AppState → pub(super); all fields pub(super) (handlers read/write ~all).
- DaemonUpdateConfig, CachedUpgradeCheck → pub(super); fields pub(super).
- ServerHandle fields → pub(super) (constructed in serve_with_options).
- DaemonConfig private fields → pub(super) (accessed by handlers).
- shared_cache_dir → pub(super) (called from serve_with_options).
These are pure visibility changes; nothing about runtime behavior moved.

SCOPE DISCIPLINE — mechanical move, zero behavior change:
- serve / serve_with_options / run / router assembly STAY in mod.rs (the
  series goal keeps mod.rs as module wiring + router assembly).
- The 14 local types AppState references (SseEvent, WsSession,
  SharedTopicState, PendingJoinResult, RestSubscription, etc.) STAY in
  mod.rs — state.rs imports them via super:: (child modules name parent
  privates). They'll be claimed by ws.rs / sse.rs / routes/ later.
- auth.rs: `use super::AppState` → `use super::state::AppState` (1 line).
- No route, handler, signature, or logic change.

CHARACTERIZATION GUARD (byte-identical):
- x0x routes (text + JSON) diffed against tests/snapshots/server-routes.*
  → IDENTICAL (P0 snapshot unchanged).
- route_set_matches_registry + manifest_matches_registry green.
- cargo doc --no-deps builds clean (public API surface preserved).
- All 1383 lib tests pass.

Gates: fmt, clippy --all-targets --all-features -- -D warnings, cargo doc,
full lib test suite (1383 passed), registry trio, route snapshot diff —
all green.

Refs #125.
